net/tls: Improve peer certificate fetching
This change should address a possible race condition when a using
application under heavy load attempts to view the DN of the peer's
(client's) certificate.

Once the handshake is successful, the function
SSL_get1_peer_certificate() should be used to obtain a copy of the
peer's certificate.  The verify_callback() mechanism is still used to
hold the last seen certificate that is being verified.  This is
necessary for error handling in order to inform a user application of
which certificate failed verification.

Customers have witnessed authorization failures because sometimes the
wrong certificate was returned purporting to be the client's
certificate.  This was traced to a possible race condition between
fetching the certificate and the callback function holding the wrong
certificate at the time.

Signed-off-by: Michael Boquard <michael@redpanda.com>